refactor: split native poll command helpers
